feat: Queue correctly returns value
This version of the queue does not handle pointers correctly,
it stores and returns a _new_ pointer to the element.

This implies it has actually created a new copy of the element,
or rather, dereferencing the element to Push() it has created
a new copy of the element. Is this a problem?

It could be.